requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
Faker==6.4.1
idna==2.10
kombu==5.0.2
prompt-toolkit==3.0.16
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
